+++ src/moves/extract_from_stack.c
@@ -6,16 +6,16 @@
 /*   By: ehosta <ehosta@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/23 01:02:44 by ehosta            #+#    #+#             */
-/*   Updated: 2025/01/24 14:51:09 by ehosta           ###   ########.fr       */
+/*   Updated: 2025/01/27 11:34:12 by ehosta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/push_swap.h"
 
-int	extract_from_stack(int *tab, size_t *head, size_t *size)
+int	extract_from_stack(int *tab, int *head, int *size)
 {
 	int		elt;
-	size_t	i;
+	int	i;
 
 	elt = tab[*head];
 	if (*head + 1 == *size)
