+++ src/sorting/is_sorted.c
@@ -6,7 +6,7 @@
 /*   By: ehosta <ehosta@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/24 17:27:35 by ehosta            #+#    #+#             */
-/*   Updated: 2025/01/24 17:30:04 by ehosta           ###   ########.fr       */
+/*   Updated: 2025/01/27 11:34:14 by ehosta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	is_sorted(t_env *env)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 
 	if (env->a_size != env->stack_size)
 		return (0);
